API

ICARE.ICARE — Module

Module ICARE

Module related to retrieving data from the AERIS/ICARE server (http://www.icare.univ-lille1.fr/). Currently, the only exported function is ftp_download to retrieve missing CALIOP data files in a specified time frame.

ICARE.ConnectionError — Type

ConnectionError

Thrown with a message as parameter, when no connection to the server can be established.

ICARE.Counter — Type

struct Counter

Stores a counter for downloads, conversions, and skipped data files as well as failed downloads. By default, all counter parts are initialised with zero, but can be adjusted by positional or individually by keyword arguments during instantiation.

ICARE.Counter — Method

Counter(;downloads::Int=0, conversions::Int=0, skipped::Int=0, failed::Int=0) -> Counter

Constructor for Counter with default init values 0 for all sub-counters. Other values can be initialised with keyword arguments for the respective sub-counter.

ICARE.File — Type

struct File

Stores relevant data about file name, extension, and paths on the local and remote system.

ICARE.File — Method

File(
    icare::SFTP.Client,
    inventory::OrderedDict,
    date::Date,
    name::String,
    convert::Bool
) -> File

Constructor for a File struct with fields for the file name, extension, location (including target, download, and remote location) and directory (dst and src); location includes the file name, dir does not. The target extensions are based on the option to convert files to a new format.

ICARE.check_localroot! — Method

check_localroot!(
    inventory::OrderedDict,
    root::AbstractString,
    product::AbstractString
)

Check, if the root has changed and update the root path and the path to the product main folder in the inventory.

ICARE.clear_dates! — Method

clear_dates!(inventory::OrderedDict)

Clear all data for dates in the inventory, but save the converted file sizes in a temp entry.

ICARE.confirm — Method

confirm(clean::Bool, logger::Logging.ConsoleLogger) -> Bool

Confirm to delete HDF4 files, if clean is set to true with terminal input and log events to logger. Return true, if clean is false, or clean is true and the user confirmed with "yes". Return false, if clean is true and user did not confirm with "yes".

ICARE.convert! — Method

convert!(
    inventory::OrderedDict,
    file::File,
    convert::Bool,
    logged::Logging.ConsoleLogger
)

Convert the file to a new file format as defined in the inventory unless file is already up-to-date or it was opted out to convert the file. Log events to logger.

ICARE.convert_file — Function

convert_file(input::String, output::String[, convert::Bool])

Convert the input to the output format unless convert is false (if provided). Both input and output can be absolute or relative paths.

ICARE.convert_hdffile! — Method

convert_hdffile!(
    inventory::OrderedDict,
    file::String,
    path::String,
    root::String,
    error_code::Int8,
    clean::Bool,
    logger::Logging.ConsoleLogger;
    h4ext::String=".hdf",
    h5ext::String=".h5"
)

Convert the file in the given path of the root directory from HDF4 to HDF5 standard (using h4ext and h5ext file extensions, respectively) and update the inventory by the file size of the new h5 file. Handle conversions and log events to logger based on the given error_code. Delete HDF4 after successful conversion, if clean is set to true.

ICARE.convert_hdffiles — Method

convert_hdffiles(
    h4files::Vector{String},
    h5files::Vector{String},
    logger::Logging.ConsoleLogger,
    logio::IOStream
)

Convert h4files to h5files and log successful conversions to logger. Successes are immediately flushed to logio to always have an up-to-date

ICARE.convertdates — Method

convertdates(startdate::Int, enddate::Int) -> startdate::Date, enddate::Date

Convert startdate and enddate from Int to Date. If the day and/or month in startdate are missing, startdate is completed with the earliest possible date (day = 01 and month = 01). If the day and/or month in enddate are missing, enddate is completed with the latest possible date (month = 12 and day = last day of that month).

ICARE.converted! — Method

converted!(inventory::OrderedDict, file::File, convert::Bool) -> Bool

Check, whether the size of the converted file is known in the inventory and matches the actual file size. Return also true, if it was opted out to convert the file.

ICARE.credentials — Method

credentials(user::String, password::String) -> Bool

Return true, if the user name and password for the ICARE server are set, otherwise false. Credentials are not checked for validity, only that they are not empty.

ICARE.data_gaps! — Method

data_gaps!(inventory::OrderedDict, dates::Vector{Date})

Add missing data gaps to the inventory from gaps in dates.

ICARE.display_gaps — Method

display_gaps(
    inventory::OrderedDict,
    daterange::@NamedTuple{start::Date,stop::Date},
    logger::Logging.ConsoleLogger
)

Log data gaps in the inventory in the given daterange to logger. Combine single dates in the inventory metadata to date ranges.

ICARE.download — Method

download!(
    icare::SFTP.Client,
    inventory::OrderedDict,
    file::File,
    update
)

Download the file from the icare server and update the inventory. Only download newer files on the server, if update is set to true.

ICARE.downloaded — Function

downloaded(
    inventory::OrderedDict,
    file::File,
    update::Bool,
    orig::Bool=false
) -> Bool

Check, whether the file.target has already been downloaded from the server to the local directory by comparing it to the filestats of the remote server in the inventory. If update is set, downloaded returns false, if newer versions of the file exist on the server. When orig is set to true, downloaded checks against the downloaded instead of the target file.

ICARE.filter_years! — Method

filter_years!(
    inventory::OrderedDict,
    years::Vector{Int},
    daterange::@NamedTuple{start::Date,stop::Date},
    resync::Bool,
    logger::Logging.ConsoleLogger
)

Filter years to keep only years within the daterange and outside the known date range of the inventory unless resync is set to true. In this case, empty dates and reset metadata. Log events to logger.

ICARE.find_conflicts — Method

find_conflicts(
    h5files::Vector{String},
    logger::Logging.ConsoleLogger
) -> Vector{Int}

Search for existing h5files and return a vector of indices of the existing files in h5files.

ICARE.find_hdffiles — Method

find_hdffiles(
    rootdir::String,
    recursive::Bool,
    ext::Union{String,Vector{String}},
    h5ext::String
) -> Tuple{Vector{String},Vector{String}}

Search for HDF4 files with the specified extension(s) in the rootdir and any subdirectories, if recursive is set to true.

Return vectors with the absolute paths and file names of the HDF4 files and the respective HDF5 counterparts exchanging the file extension to h5ext.

ICARE.h4cleanup — Method

h4cleanup(h4files::Vector{String}, clean::Bool, logger::Logging.ConsoleLogger)

Delete all h4files, if clean is set to true and log deletions to logger.

ICARE.h5upgrade! — Function

h5upgrade!(
  inventory::OrderedDict,
  date::Date,
  h4file::String,
  h5file::String="";
  force::Bool=false
)

Upgrade the h4file to h5file for the given date and add the size of the new HDF5 file to the inventory. Overwrite any existing h5file, if force is set to true.

ICARE.hdfupgrade — Function

hdfupgrade(
    product::String,
    version::Float64 = 4.51;
    localroot::String = ".",
    remoteroot::String = "/SPACEBORNE/CALIOP",
    user::String = "",
    password::String = "",
    clean::Bool = false,
    h4ext::String = ".hdf",
    h5ext::String = ".h5",
    logfile::String = "conversions.log",
    loglevel::Symbol = :Debug
)

Convert all HDF4 data files to HDF5 files (with h4ext/h5ext file extensions, respectively) that are part of specified product in the given version.

To check, whether hdf files are part of the product, the are compared to the inventory file in the product folder in the localroot. Additional checks against the ICARE server can be performed with a significant increase in processing time, if the ICARE user name and password are given as well as the remoteroot of the product folder on the server.

By default, files are upgraded to HDF5 standard and original files are kept. If clean is set to true, the original files are deleted upon confirmation. This means that data files will have to be re-downloaded, if data files in the original HDF4 format are needed!

Events of the desired loglevel will be logged to logfile. A timestamp will be added to each logfile so that a unique logfile is created for every run and logfiles are not overwritten.

ICARE.icare_connect — Function

icare_connect(
    user::String,
    password::String,
    root::String,
    product::String
) -> SFTP.Client

Securely connect to the server with SFTP using the credentials user and password and changing to the product folder in the root directory.

Several checks are performed about the connection and folder structure and a SFTP.Client type with all the relevant information about the server is returned.

ICARE.in_database — Method

in_database(
    inventory::OrderedDict,
    icare::SFTP.Client,
    root::String,
    path::String,
    file::String,
    h5files::Vector{String}
) -> Int8

Determine, whether the file in the given path of the root directory is part of the inventory. Check the icare server, if files for a given date are unavailable. Check against available h5files in the same folder for already converted files. Return an error code for possible events:

-1: corrupt inventory (should not happen)

0: file in database

1: h5 already exists; conversion skipped

2: existing h5 file/inventory need update

3: not in local inventory

4: file not on server

5: no data reported for date (should already be caught elsewhere)

6: date outside expected date range (should already be caught elsewhere)

ICARE.init_logging — Method

init_logging(logfile::String, rootdir::String, loglevel::Symbol) -> Tuple{String,Logging.LogLevel}

Add a timestamp to the logfile. If no path is given in the file name, save logfile to rootdir. Return the updated logfile and the loglevel as Logging.LogLevel.

ICARE.load_database — Method

load_database(
  productfolder::String,
  localroot::String,
  remoteroot::String,
  user::String,
  password::String,
  logger::Logging.ConsoleLogger
) -> Tuple{Union{SFTP.Client,Nothing},OrderedDict}

Load the inventory in the given productfolder of the localroot. Additionally, access ICARE server data, if the user name, password, and remoteroot are given. One or the other must be given or an IOError will be thrown. All events are logged to logger.

ICARE.load_inventory! — Method

load_inventory!(inventory::OrderedDict, file::AbstractString)

Load data from a yaml file to the inventory.

ICARE.log_counter — Method

log_counter(counter::Counter, logger::Logging.ConsoleLogger, t0::DateTime)

Log the number of downloaded, skipped, and converted files saved in counter to logger together with the time it took since t0.

ICARE.new_date! — Method

new_date!(inventory::OrderedDict, date::Date) -> Bool

Add the given date to the inventory, if missing. Return true, if the date was added, otherwise false.

ICARE.new_inventory! — Method

new_inventory!(
    icare::SFTP.Client,
    inventory::OrderedDict,
    root::String,
    product::String,
    convert::Bool,
    logger::Logging.ConsoleLogger
)

Initialise a new and empty inventory.

ICARE.newext — Function

newext([inventory::OrderedDict, convert::Bool]) -> String

Return the extension of the converted file format or the original file extension from the inventory, when convert is false (no conversion). If no arguments are provided, return the default extension of the converted file format.

ICARE.newext! — Method

newext!(inventory::OrderedDict, convert::Bool)

Check and update the converted file extension in the inventory. If convert is false, the target extension is set to the original file extension.

ICARE.product_database! — Method

product_database!(
    icare::SFTP.Client,
    inventory::OrderedDict,
    root::String,
    product::String,
    daterange::@NamedTuple{start::Date,stop::Date},
    convert::Bool,
    resync::Bool,
    logger::Logging.ConsoleLogger
)

Initiate the inventory of icare server-side data files for the product in the remoteroot directory. Either read the database from the yaml file in the product folder or initialise a new database. If the daterange of the selected dates is (partly) outside the inventory date range, the inventory is updated for these extended periods. The whole inventory can be updated by setting resync to true. Additional checks are performed, whether the root folder was moved. In that case, the inventory is updated and a warning is issued. The target file extension for converted files is set based on the convert option. Updates are logged to the screen and the log file with logger.

ICARE.remotefiles! — Method

remotefiles!(
    icare::SFTP.Client,
    inventory::OrderedDict,
    date::Date
) -> Bool

Add file stats for all granules of the date based on the icare server data. File stats are only added for dates with no file data. Indicate updates in the inventory by the returned Bool.

ICARE.rmh4 — Method

rmh4(
    file::String,
    clean::Bool,
    logger::Logging.ConsoleLogger
)

Delete the HDF4 file, if clean is set to true and log the event to logger.

ICARE.save_inventory — Method

save_inventory(inventory::OrderedDict, t::DateTime)

Save the inventory to <product path>/.inventory.yaml if changes occurred since time t.

ICARE.set_converted_size! — Method

set_converted_size!(
    inventory::OrderedDict,
    file::File,
    convert::Bool,
    logger::Logging.ConsoleLogger
)

Set the size of the converted file in the inventory and mark the inventory as updated. Log events to logger.

ICARE.set_localroot — Method

set_localroot(localroot::String, mainfolder::String) -> String

Define the product folder on the local system from the localroot and the mainfolder containing all the year folders for the ICARE data.

ICARE.setup_conversion! — Method

setup_conversion!(
    h4files::Vector{String},
    h5files::Vector{String},
    h5::Vector{Int},
    h5exists::Symbol,
    logger::Logging.ConsoleLogger
) -> Vector{String}

Adjust the list of existing h4files and target h5files using the indexes of existing h5 files based on the choices, if h5exists. Conversions with existing h5 files can either be skipped or cancelled on the first encounter (deleting the respective entries in h4files and h5files), overwritten or aborted (emptying the h5files vector). Choices and number of conversions are logged to logger.

ICARE.sftp_download — Function

sftp_download(
    user::String,
    password::String,
    product::String,
    startdate::Int,
    enddate::Int=-1;
    version::Union{Nothing,Real} = 4.51,
    remoteroot::String = "/SPACEBORNE/CALIOP/",
    localroot::String = ".",
    convert::Bool = true,
    resync::Bool = false,
    update::Bool = false,
    logfile::String = "downloads.log",
    loglevel::Symbol = :Debug

)

Download satellite data from the Aeris/ICARE server.

Note

To use sftp_download, an Aeris/ICARE account is needed that is available for free for non-commercial use.

Positional arguments

user::String/password::String: Aeris/ICARE account credentials

product::String: The desired product to download (matches the folder name excluding the version number, e.g., 05kmCPro)

startdate::Int/enddate::Int: The start/end date for the download period as Int (format: yyyy[mm[dd]]);

In the dates, the day and month part can be omitted. In this case, the earliest possible start date is selected and the latest possible end date, e.g. 202003 will give a start date of 2020-03-01 and an end date of 2020-03-31. The end date is optional, if omitted, the period defined by startdate is downloaded, either a day, or a month (if the day part is omitted) or a year (if both day and month are omitted).

Keyword arguments

version::Union{Nothing,Real}: The version number of the product (default: 4.51).

remoteroot::String: The root path on the remote server (default: "/SPACEBORNE/CALIOP/").

localroot::String: The root path on the local machine containing the product folder (default: ".").

convert::Bool: Whether or not to convert the downloaded files to another file format (default: true).

resync::Bool: Whether to re-synchronize the local inventory with the remote server (default: false).

update::Bool: Whether to update the local files if newer versions are available on the remote server (default: false).

logfile::String: The name of the log file (default: "downloads.log"; the name will be appended by the current date and time).

loglevel::Symbol: The log level for the download process (default: :Debug).

Warning

The update option automatically resynchronises the inventory as well.

Re-synchronisation of the inventory will take several minutes up to hours!

For custom version formats, the version can be set to nothing and included in the product string. By default, the product folder is constructed as <product>.v<X.XX> with the version as float with two decimal places independent of the input format.

By default, hdf files (version 4) are assumed as download source, which will be converted to .h5 (HDF5) file unless convert is set to false.

ICARE.sync! — Method

sync!(
    icare::SFTP.Client,
    inventory::OrderedDict,
    daterange::@NamedTuple{start::Date, stop::Date},
    convert::Bool,
    update::Bool,
    resync::Bool,
    logger::Logging.ConsoleLogger,
    logio::IO,
    counter::Counter
)

Synchronize the files for the selected daterange from the icare server with the local system. If set, update and convert files to the latest version available on the server and a predefined file format. Dates and files are compared to the inventory and the inventory is updated, if necessary. Increase the respective counter for each sync action and log logger events to a log file in the logio I/O stream.

ICARE.sync_database! — Method

sync_database!(
    icare::SFTP.Client,
    inventory::OrderedDict,
    years::Vector{Int},
    daterange::@NamedTuple{start::Date,stop::Date},
    convert::Bool,
    logger::Logging.ConsoleLogger
)

Sync the inventory with the icare server for the given daterange and years. Consider conversion to a new file format based on the convert option. Log events to logger.

ICARE.update_stats! — Method

update_stats!(
    icare::SFTP.Client,
    inventory::OrderedDict,
    file::File,
    resync::Bool,
    logger::Logging.ConsoleLogger
)

Update the file stats in the inventory with the remote icare server. If the size of the converted file does not match the inventory, it is updated, too. Ignore already checked dates, if the inventory was already resynced. Reset data for files previously in the database, but currently not available on the server. Log events to logger.

ICARE.valid_dir — Method

valid_dir(
    path::String,
    inventory::OrderedDict,
    logger::Logging.ConsoleLogger
) -> Bool

Check that the path is a valid directory within the ICARE folder structure. Return true, if the folder meets the ICARE standard, false otherwise. Log events to logger (using inventory data if needed).