Analysing and manipulating the inventory

Analysing the inventory

The structure and files including file stats are saved in an inventory that is returned by all API functions sftp_download, clean, convert!, and convert. The inventory is a SortedDict that saves relevant data such as the available data range, the file count and the overall size in the metadata. Stats about the data file are kept in the "dates" section and missing dates are listed in "gaps".

Manipulating the inventory

Several functions exist to help shaping the database to the users need.

Batch conversions

Methods convert/convert! can be used to convert data files to a new file format. By default, HDF4 files are upgraded to HDF5. The same could be achieved by re-running sftp_download with convert option set to true. If the original files are already downloaded, then sftp_download will not re-download these files and just convert them. If you have already loaded the inventory, then convert! is the more performant option; convert is a convenience method that will load the inventory first and then call convert!.

Base.convert — Method

convert(
    root::String=".";
    sizecheck::Bool=false,
    logfile::String = "conversions.log",
    loglevel::Symbol = :Debug
) -> SortedDict{String, Any}

Convert all files in root and part of the inventory to a new format as defined in the inventory. If files of the new format already exist, they are skipped unless sizecheck is set to true, which will reconvert any file whose size differs from that listed in the inventory. Logging is written to logfile with the specified loglevel. A timestamp is added to the log file name to avoid overwriting existing logs. The function returns the updated inventory.

ICARE.convert! — Function

convert!(
    inventory::SortedDict{String, Any};
    sizecheck::Bool=false,
    logfile::String = "conversions.log",
    loglevel::Symbol = :Debug
) -> SortedDict{String, Any}

Convert all files in the local database and part of the inventory to a new format as defined in the inventory. If files of the new format already exist, they are skipped unless sizecheck is set to true, which will reconvert any file whose size differs from that listed in the inventory. Logging is written to logfile with the specified loglevel. A timestamp is added to the log file name to avoid overwriting existing logs. The function returns the updated inventory.

Cleaning up the inventory

Function clean can be used to remove all files in a product folder that do not belong to the database, i.e. that don't exist on the ICARE server. Files will be permanently deleted, but the user will always be shown, which files will be deleted, and ask for confirmation.

By passing a file extension or a vector of file extensions to the keyword argument keepext, those files will be ignored for the clean-up. This can be useful, for example, to keep log files defaulting to the .log extension. Otherwise, they will be removed not counting as part of the database and only the log file from the current cleaning run is saved.

The database itself can be cleaned up as well. By default, both the original and converted files are considered part of the database. But one or the other file type can be removed during cleaning. This is done by specifying which datatype shall be removed to the erase keyword argument with the Extension enum. The two choices available are original and converted. The enums are passed directly to the erase kwarg (without quotes or any other specifier). They are constants in ICARE that are exported from the package. Therefore, they should not be overwritten.

Warning

The Extension enum together with the values original and converted is exported by ICARE.jl. Do not unintentionally overwrite them or you will get error messages, when trying to set the erase option in clean.

Danger

If the original files are removed during clean, they are permanently lost and will have to be re-downloaded, if needed. If converted files are deleted, they can be converted again from the original, which is much faster than downloading.

ICARE.clean — Function

clean(
    root::AbstractString=".";
    keepext::Union{AbstractString,Vector{<:AbstractString}}="",
    erase::Extension=none,
    logfile::String = "clean.log",
    loglevel::Symbol = :Debug
) -> SortedDict{String, Any}

clean(
    inventory::SortedDict{String, Any};
    keepext::Union{AbstractString,Vector{<:AbstractString}}="",
    erase::Extension=none,
    logfile::String = "clean.log",
    loglevel::Symbol = :Debug
) -> SortedDict{String, Any}

Clean a product folder recursively from all content not listed in the inventory, i.e. not available on the ICARE server. The function has two methods – you can either provide an AbstractString with the path of the product folder or the inventory as SortedDict of the product. Both methods return the inventory for reference.

Keyword Arguments

keepext::Union{AbstractString,Vector{<:AbstractString}}: One or multiple (as vector) file extensions (e.g. ".log", [".yaml", ".log"]) to keep during clean-up even if not part of the inventory. Can be used to keep log or metadata files.

erase::Extension: Allows to clean up the database itself. You can choose to erase original files (defined by "ext" in the inventory metadata), converted files (defined by "newext" in the inventory metadata) or none of the file types.

logfile::String: The name of the log file (default: "clean.log"; the name will be appended by the current date and time).

loglevel::Symbol: The log level for the download process (default: :Debug).

ICARE.Extension — Type

Enum Extension

Store general choices for available file extensions in the database.